#11: Separated GlobalSettingsController and its companion object
